Adds BatteryMonitor and PushNotifications services to check message battery levels.
In this commit, we add two major classes. The first of which is PushNotifications.ts, which exports
a service that will send push notifications on request. The second of which is BatteryMonitor.ts which
subscribes to the new events from the MqttHandler message streams and will send a push notification if
the battery level is lower than a given threshold.