Successfully renders all of the battery trends within the same chart.
In this commit, we fix a bug where all events wold appear as the same trend looping
back within the graph. I believe this was being caused by an erroneous line of
javascript where all of the datasets contained the exact same values.

I am still not sure whether the graph is updating to new data yet.