Redesigns project structure to utilise ObserverSubject's.
In this commit we redesign the PushNotifications.ts and SettingsManager.ts to
utilise the Observer subject pattern. This means that our components can be
dynamic when the service file changes its own state. We also fix parsing of the
settings state from local storage. This allows us to dynamically see the
notifications state within the application.